Keep proxy metadata middleware with provider
